+++ cmd/argo/commands/uninstall.go
@@ -3,7 +3,7 @@ package commands
 import (
 	"fmt"
 
-	wfv1 "github.com/argoproj/argo/pkg/apis/workflow/v1alpha1"
+	"github.com/argoproj/argo/pkg/apis/workflow"
 	"github.com/argoproj/argo/workflow/common"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
@@ -77,21 +77,6 @@ func uninstall(cmd *cobra.Command, args []string) {
 		fmt.Printf("ConfigMap '%s' deleted\n", uninstallArgs.configMap)
 	}
 
-	// Delete the workflow CRD
-	apiextensionsclientset, err := apiextensionsclient.NewForConfig(restConfig)
-	if err != nil {
-		log.Fatalf("%+v", err)
-	}
-	err = common.DeleteCustomResourceDefinition(apiextensionsclientset)
-	if err != nil {
-		if !apierr.IsNotFound(err) {
-			log.Fatalf("Failed to delete CustomResourceDefinition '%s': %v", wfv1.CRDFullName, err)
-		}
-		fmt.Printf("CustomResourceDefinition '%s' not found\n", wfv1.CRDFullName)
-	} else {
-		fmt.Printf("CustomResourceDefinition '%s' deleted\n", wfv1.CRDFullName)
-	}
-
 	// Delete role binding
 	if err := clientset.RbacV1beta1().ClusterRoleBindings().Delete(ArgoClusterRole, &metav1.DeleteOptions{}); err != nil {
 		if !apierr.IsNotFound(err) {
@@ -110,5 +95,18 @@ func uninstall(cmd *cobra.Command, args []string) {
 		fmt.Printf("ServiceAccount '%s' in namespace '%s' not found\n", ArgoServiceAccount, uninstallArgs.namespace)
 	} else {
 		fmt.Printf("ServiceAccount '%s' deleted\n", ArgoServiceAccount)
+	}
+
+	// Delete the workflow CRD
+	apiextensionsclientset := apiextensionsclient.NewForConfigOrDie(restConfig)
+	crdClient := apiextensionsclientset.Apiextensions().CustomResourceDefinitions()
+	err = crdClient.Delete(workflow.FullName, nil)
+	if err != nil {
+		if !apierr.IsNotFound(err) {
+			log.Fatalf("Failed to delete CustomResourceDefinition '%s': %v", workflow.FullName, err)
+		}
+		fmt.Printf("CustomResourceDefinition '%s' not found\n", workflow.FullName)
+	} else {
+		fmt.Printf("CustomResourceDefinition '%s' deleted\n", workflow.FullName)
 	}
 }