+++ pkg/apis/workflow/v1alpha1/types.go
@@ -7,18 +7,6 @@ import (
 
 	apiv1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-)
-
-// CRD constants
-const (
-	CRDKind      string = "Workflow"
-	CRDSingular  string = "workflow"
-	CRDPlural    string = "workflows"
-	CRDShortName string = "wf"
-	CRDGroup     string = "argoproj.io"
-	CRDVersion   string = "v1alpha1"
-	CRDFullName  string = CRDPlural + "." + CRDGroup
 )
 
 // TemplateType is the type of a template
@@ -45,16 +33,13 @@ const (
 	NodeError     NodePhase = "Error"
 )
 
-// Create a Rest client with the new CRD Schema
-var SchemeGroupVersion = schema.GroupVersion{Group: CRDGroup, Version: CRDVersion}
-
 // +genclient
 // +genclient:noStatus
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
 
 // Workflow is the definition of our CRD Workflow class
 type Workflow struct {
-	metav1.TypeMeta   `json:",inline,squash"`
+	metav1.TypeMeta   `json:",inline"`
 	metav1.ObjectMeta `json:"metadata"`
 	Spec              WorkflowSpec   `json:"spec"`
 	Status            WorkflowStatus `json:"status"`
@@ -64,7 +49,7 @@ type Workflow struct {
 
 // WorkflowList is list of Workflow resources
 type WorkflowList struct {
-	metav1.TypeMeta `json:",inline,squash"`
+	metav1.TypeMeta `json:",inline"`
 	metav1.ListMeta `json:"metadata"`
 	Items           []Workflow `json:"items"`
 }
@@ -119,7 +104,7 @@ type Template struct {
 	// Deamon will allow a workflow to proceed to the next step so long as the container reaches readiness
 	Daemon *bool `json:"daemon,omitempty"`
 
-	// Step
+	// Steps define a series of sequential/parallal workflow steps
 	Steps [][]WorkflowStep `json:"steps,omitempty"`
 
 	// Container
@@ -147,6 +132,8 @@ type Template struct {
 	// before the system actively tries to terminate the pod; value must be positive integer
 	// This field is only applicable to container and script templates.
 	ActiveDeadlineSeconds *int64 `json:"activeDeadlineSeconds,omitempty"`
+
+	RetryStrategy *RetryStrategy `json:"retryStrategy,omitempty"`
 }
 
 // Inputs are the mechanism for passing parameters, artifacts, volumes from one template to another
@@ -157,12 +144,34 @@ type Inputs struct {
 
 // Parameter indicate a passed string parameter to a service template with an optional default value
 type Parameter struct {
-	Name    string  `json:"name"`
-	Value   *string `json:"value,omitempty"`
+	// Name is the parameter name
+	Name string `json:"name"`
+
+	// Default is the default value to use for an input parameter if a value was not supplied
 	Default *string `json:"default,omitempty"`
 
-	// Path describes the location in which to retrieve the output parameter value from
+	// Value is the literal value to use for the parameter.
+	// If specified in the context of an input parameter, the value takes precedence over any passed values
+	Value *string `json:"value,omitempty"`
+
+	// ValueFrom is the source for the output parameter's value
+	ValueFrom *ValueFrom `json:"valueFrom,omitempty"`
+}
+
+// ValueFrom describes a location in which to obtain the value to a parameter
+type ValueFrom struct {
+	// Path in the container to retrieve an output parameter value from in container templates
 	Path string `json:"path,omitempty"`
+
+	// JSONPath of a resource to retrieve an output parameter value from in resource templates
+	JSONPath string `json:"jsonPath,omitempty"`
+
+	// JQFilter expression against the resource object in resource templates
+	JQFilter string `json:"jqFilter,omitempty"`
+
+	// Parameter reference to a step or dag task in which to retrieve an output parameter value from
+	// (e.g. '{{steps.mystep.outputs.myparam}}')
+	Parameter string `json:"parameter,omitempty"`
 }
 
 // Artifact indicates an artifact to place at a specified path
@@ -180,7 +189,8 @@ type Artifact struct {
 	// From allows an artifact to reference an artifact from a previous step
 	From string `json:"from,omitempty"`
 
-	ArtifactLocation `json:",inline,squash"`
+	// ArtifactLocation contains the location of the artifact
+	ArtifactLocation `json:",inline"`
 }
 
 // ArtifactLocation describes a location for a single or multiple artifacts.
@@ -194,6 +204,7 @@ type ArtifactLocation struct {
 	Artifactory *ArtifactoryArtifact `json:"artifactory,omitempty"`
 }
 
+// Outputs hold parameters, artifacts, and results from a step
 type Outputs struct {
 	// Parameters holds the list of output parameters produced by a step
 	Parameters []Parameter `json:"parameters,omitempty"`
@@ -226,9 +237,9 @@ type Arguments struct {
 
 // Sidecar is a container which runs alongside the main container
 type Sidecar struct {
-	apiv1.Container `json:",inline,squash"`
+	apiv1.Container `json:",inline"`
 
-	SidecarOptions `json:",inline,squash"`
+	SidecarOptions `json:",inline"`
 }
 
 // SidecarOptions provide a way to customize the behavior of a sidecar and how it
@@ -267,6 +278,21 @@ type WorkflowStatus struct {
 	PersistentVolumeClaims []apiv1.Volume `json:"persistentVolumeClaims,omitempty"`
 }
 
+// GetNodesWithRetries returns a list of nodes that have retries.
+func (wfs *WorkflowStatus) GetNodesWithRetries() []NodeStatus {
+	var nodesWithRetries []NodeStatus
+	for _, node := range wfs.Nodes {
+		if node.RetryStrategy != nil {
+			nodesWithRetries = append(nodesWithRetries, node)
+		}
+	}
+	return nodesWithRetries
+}
+
+type RetryStrategy struct {
+	Limit *int32 `json:"limit,omitempty"`
+}
+
 type NodeStatus struct {
 	// ID is a unique identifier of a node within the worklow
 	// It is implemented as a hash of the node name, which makes the ID deterministic
@@ -297,6 +323,8 @@ type NodeStatus struct {
 
 	// Daemoned tracks whether or not this node was daemoned and need to be terminated
 	Daemoned *bool `json:"daemoned,omitempty"`
+
+	RetryStrategy *RetryStrategy `json:"retryStrategy,omitempty"`
 
 	// Outputs captures output parameter values and artifact locations
 	Outputs *Outputs `json:"outputs,omitempty"`
@@ -344,6 +372,12 @@ func (n NodeStatus) Successful() bool {
 	return n.Phase == NodeSucceeded || n.Phase == NodeSkipped
 }
 
+// CanRetry returns whether the node should be retried or not.
+func (n NodeStatus) CanRetry() bool {
+	// TODO(shri): Check if there are some 'unretryable' errors.
+	return n.Completed() && !n.Successful()
+}
+
 type S3Bucket struct {
 	Endpoint        string                  `json:"endpoint"`
 	Bucket          string                  `json:"bucket"`
@@ -354,7 +388,7 @@ type S3Bucket struct {
 }
 
 type S3Artifact struct {
-	S3Bucket `json:",inline,squash"`
+	S3Bucket `json:",inline"`
 	Key      string `json:"key"`
 }
 
@@ -371,7 +405,7 @@ type ArtifactoryAuth struct {
 }
 
 type ArtifactoryArtifact struct {
-	ArtifactoryAuth `json:",inline,squash"`
+	ArtifactoryAuth `json:",inline"`
 	URL             string `json:"url"`
 }