+++ workflow/executor/resource.go
@@ -1,0 +1,188 @@
+package executor
+
+import (
+	"bufio"
+	"bytes"
+	"fmt"
+	"os/exec"
+	"strings"
+
+	"github.com/argoproj/argo/errors"
+	log "github.com/sirupsen/logrus"
+	"github.com/tidwall/gjson"
+	"k8s.io/apimachinery/pkg/labels"
+)
+
+// ExecResource will run kubectl action against a manifest
+func (we *WorkflowExecutor) ExecResource(action string, manifestPath string) (string, error) {
+	args := []string{
+		action,
+	}
+	if action == "delete" {
+		args = append(args, "--ignore-not-found")
+	}
+	args = append(args, "-f")
+	args = append(args, manifestPath)
+	args = append(args, "-o")
+	args = append(args, "name")
+	cmd := exec.Command("kubectl", args...)
+	log.Info(strings.Join(cmd.Args, " "))
+	out, err := cmd.Output()
+	if err != nil {
+		exErr := err.(*exec.ExitError)
+		errMsg := strings.TrimSpace(string(exErr.Stderr))
+		return "", errors.New(errors.CodeBadRequest, errMsg)
+	}
+	resourceName := strings.TrimSpace(string(out))
+	log.Infof(resourceName)
+	return resourceName, nil
+}
+
+// gjsonLabels is an implementation of labels.Labels interface
+// which allows us to take advantage of k8s labels library
+// for the purposes of evaluating fail and success conditions
+type gjsonLabels struct {
+	json []byte
+}
+
+// Has returns whether the provided label exists.
+func (g gjsonLabels) Has(label string) bool {
+	return gjson.GetBytes(g.json, label).Exists()
+}
+
+// Get returns the value for the provided label.
+func (g gjsonLabels) Get(label string) string {
+	return gjson.GetBytes(g.json, label).String()
+}
+
+// WaitResource waits for a specific resource to satisfy either the success or failure condition
+func (we *WorkflowExecutor) WaitResource(resourceName string) error {
+	if we.Template.Resource.SuccessCondition == "" && we.Template.Resource.FailureCondition == "" {
+		return nil
+	}
+	var successReqs labels.Requirements
+	if we.Template.Resource.SuccessCondition != "" {
+		successSelector, err := labels.Parse(we.Template.Resource.SuccessCondition)
+		if err != nil {
+			return errors.Errorf(errors.CodeBadRequest, "success condition '%s' failed to parse: %v", we.Template.Resource.SuccessCondition, err)
+		}
+		log.Infof("Waiting for conditions: %s", successSelector)
+		successReqs, _ = successSelector.Requirements()
+	}
+
+	var failReqs labels.Requirements
+	if we.Template.Resource.FailureCondition != "" {
+		failSelector, err := labels.Parse(we.Template.Resource.FailureCondition)
+		if err != nil {
+			return errors.Errorf(errors.CodeBadRequest, "fail condition '%s' failed to parse: %v", we.Template.Resource.FailureCondition, err)
+		}
+		log.Infof("Failing for conditions: %s", failSelector)
+		failReqs, _ = failSelector.Requirements()
+	}
+
+	cmd := exec.Command("kubectl", "get", resourceName, "-w", "-o", "json")
+	stdout, err := cmd.StdoutPipe()
+	if err != nil {
+		return errors.InternalWrapError(err)
+	}
+	reader := bufio.NewReader(stdout)
+	log.Info(strings.Join(cmd.Args, " "))
+	if err := cmd.Start(); err != nil {
+		return errors.InternalWrapError(err)
+	}
+	defer func() {
+		_ = cmd.Process.Kill()
+	}()
+	for {
+		jsonBytes, err := readJSON(reader)
+		if err != nil {
+			// TODO: it's possible that kubectl might EOF upon an API disconnect.
+			// In this case, we would want to refork kubectl again.
+			return errors.InternalWrapError(err)
+		}
+		log.Info(string(jsonBytes))
+		ls := gjsonLabels{json: jsonBytes}
+		for _, req := range failReqs {
+			failed := req.Matches(ls)
+			msg := fmt.Sprintf("failure condition '%s' evaluated %v", req, failed)
+			log.Infof(msg)
+			if failed {
+				// TODO: need a better error code instead of BadRequest
+				return errors.Errorf(errors.CodeBadRequest, msg)
+			}
+		}
+		numMatched := 0
+		for _, req := range successReqs {
+			matched := req.Matches(ls)
+			log.Infof("success condition '%s' evaluated %v", req, matched)
+			if matched {
+				numMatched++
+			}
+		}
+		log.Infof("%d/%d success conditions matched", numMatched, len(successReqs))
+		if numMatched >= len(successReqs) {
+			break
+		}
+	}
+	return nil
+}
+
+// readJSON reads from a reader line-by-line until it reaches "}\n" indicating end of json
+func readJSON(reader *bufio.Reader) ([]byte, error) {
+	var buffer bytes.Buffer
+	for {
+		line, err := reader.ReadBytes('\n')
+		if err != nil {
+			return nil, err
+		}
+		isDelimiter := len(line) == 2 && line[0] == byte('}')
+		line = bytes.TrimSpace(line)
+		_, err = buffer.Write(line)
+		if err != nil {
+			return nil, err
+		}
+		if isDelimiter {
+			break
+		}
+	}
+	return buffer.Bytes(), nil
+}
+
+// SaveResourceParameters will save any resource output parameters
+func (we *WorkflowExecutor) SaveResourceParameters(resourceName string) error {
+	log.Infof("Saving resource output parameters")
+	if len(we.Template.Outputs.Parameters) == 0 {
+		log.Infof("No output parameters, nothing to do")
+		return nil
+	}
+	for i, param := range we.Template.Outputs.Parameters {
+		if param.ValueFrom == nil {
+			continue
+		}
+		var cmd *exec.Cmd
+		if param.ValueFrom.JSONPath != "" {
+			cmd = exec.Command("kubectl", "get", resourceName, "-o", fmt.Sprintf("jsonpath='%s'", param.ValueFrom.JSONPath))
+		} else if param.ValueFrom.JQFilter != "" {
+			cmdStr := fmt.Sprintf("kubectl get %s -o json | jq -c '%s'", resourceName, param.ValueFrom.JQFilter)
+			cmd = exec.Command("sh", "-c", cmdStr)
+		} else {
+			continue
+		}
+		log.Info(cmd.Args)
+		out, err := cmd.Output()
+		if err != nil {
+			if exErr, ok := err.(*exec.ExitError); ok {
+				log.Errorf("`%s` stderr:\n%s", cmd.Args, string(exErr.Stderr))
+			}
+			return errors.InternalWrapError(err)
+		}
+		output := string(out)
+		we.Template.Outputs.Parameters[i].Value = &output
+		log.Infof("Saved output parameter: %s, value: %s", param.Name, output)
+	}
+	err := we.AnnotateOutputs()
+	if err != nil {
+		return err
+	}
+	return nil
+}